focus on plot library
