+++ code/lib/plotting/__init__.py
@@ -1,2 +1,3 @@
-from .single_model_plots import *
-from .multiple_model_plots import *
+from .single_model_single_variation import *
+from .single_model_multi_variations import *
+from .multi_model_plots import *
