+++ scripts/experiments/essentia-prepare_search_evaluate_experiments.sh
@@ -15,6 +15,6 @@ for v in ${variable[@]}; do
 done
 
 # Compare the results of the experiments
-python code/plot_evaluation_results.py $MODEL_NAME
+python code/plot_evaluation_results_comparisons.py $MODEL_NAME
 
 #############################################################################
