+++ scripts/experiments/sinet-prepare_search_evaluate_experiments.sh
@@ -37,6 +37,6 @@ for v1 in ${variable1[@]}; do
 done
 
 # Compare the results of the experiments
-python code/plot_evaluation_results.py =$1
+python code/plot_evaluation_results_comparisons.py =$1
 
 #############################################################################
