+++ scripts/experiments/sinet-search_evaluate_experiments.sh
@@ -32,7 +32,7 @@ for file in "$EMBED_DIR/"*; do # for each embedding dir
 done
 
 # Compare the results of the experiments
-python code/plot_evaluation_results.py =$1
+python code/plot_evaluation_results_comparisons.py =$1
 
 
 #############################################################################
