+++ scripts/experiments/vggish-prepare_search_evaluate_experiments.sh
@@ -26,6 +26,6 @@ for v1 in ${variable1[@]}; do
 done
 
 # Compare the results of the experiments
-python code/plot_evaluation_results.py =audioset-vggish-3
+python code/plot_evaluation_results_comparisons.py =audioset-vggish-3
 
 #############################################################################
